+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/experience/enchantments/enchantment/ArmorPiercer.java
@@ -1,0 +1,49 @@
+package insane96mcp.iguanatweaksexpanded.module.experience.enchantments.enchantment;
+
+import insane96mcp.iguanatweaksexpanded.module.experience.enchanting.EnchantingFeature;
+import insane96mcp.iguanatweaksreborn.data.generator.ITRItemTagsProvider;
+import insane96mcp.iguanatweaksreborn.module.combat.PiercingDamage;
+import insane96mcp.iguanatweaksreborn.module.experience.enchantments.EnchantmentsFeature;
+import insane96mcp.iguanatweaksreborn.module.experience.enchantments.enchantment.damage.BonusDamageEnchantment;
+import net.minecraft.tags.TagKey;
+import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.world.entity.ai.attributes.AttributeModifier;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.enchantment.Enchantment;
+import net.minecraft.world.item.enchantment.EnchantmentCategory;
+import net.minecraftforge.event.ItemAttributeModifierEvent;
+
+import java.util.UUID;
+
+public class ArmorPiercer extends Enchantment implements IAttributeEnchantment {
+    public static final TagKey<Item> ACCEPTS_ENCHANTMENT = ITRItemTagsProvider.create("enchanting/accepts_armor_piercer_enchantments");
+    static final EnchantmentCategory CATEGORY = EnchantmentCategory.create("armor_piercer_enchantment", (item) -> {
+        return item.builtInRegistryHolder().is(ACCEPTS_ENCHANTMENT);
+    });
+
+    public static final UUID MODIFIER_UUID = UUID.fromString("74e97c20-6a62-482f-b909-e709087b066a");
+
+    public ArmorPiercer() {
+        super(Rarity.UNCOMMON, CATEGORY, new EquipmentSlot[]{EquipmentSlot.MAINHAND});
+    }
+
+    @Override
+    public int getMaxLevel() {
+        return 5;
+    }
+
+    public int getMinCost(int lvl) {
+        return 4 + (lvl - 1) * 8;
+    }
+
+    public int getMaxCost(int lvl) {
+        return this.getMinCost(lvl) + 20;
+    }
+
+    @Override
+    public void applyAttributeModifier(ItemAttributeModifierEvent event, int enchantmentLvl) {
+        if (event.getSlotType() != EquipmentSlot.MAINHAND)
+            return;
+        event.addModifier(PiercingDamage.PIERCING_DAMAGE.get(), new AttributeModifier(MODIFIER_UUID, "Armor Piercier Enchantment Modifier", 0.35d * enchantmentLvl, AttributeModifier.Operation.ADDITION));
+    }
+}